Fix new user registration via social network login
User registration via social network login was failing CSRF validation,
which causes the auth data to be purged from the session state. This
seems to be due to a host name mismatch, wherein our requests come from
a subdomain of whispersystems.org, but the app thinks its host is named
something like ec2-NN-blahblahblah...

The user-facing symptom was an error about minimum password length.

The canonical solution (according to Discourse devs) is to add a
host_names config to database.yml with our custom domain, but that turns
out to be difficult to do on Heroku. Hopefully they will externalize
this setting to some other config file that is more under our control.

As a workaround, this patch disables CSRF validation for user creation.
Since you do not have to be logged in to create a user, it is unclear to
me what CSRF is protecting against in this situation.